updating audio recording transfer scripts to not mix and compress files if the FTP server does not Ping

+++ bin/AST_CRON_mix_recordings_MP3.pl
@@ -32,6 +32,7 @@
 # 51122-1455 - Added soxmix and sox binary path check
 # 60616-1027 - Modified to convert to MP3 format
 # 60807-1308 - Modified to use /etc/astguiclient.conf for settings 
+# 71004-1124 - Changed to not move ORIG recordings if FTP server does not Ping
 # 
 
 
@@ -145,42 +146,43 @@ foreach(@FILES)
 
 		if ($v) {print "|$INfile|    |$OUTfile|     |$ALLfile|\n\n";}
 
-			`$soxmixbin "$dir1/$INfile" "$dir1/$OUTfile" "$dir1/$ALLfile"`;
-		if ($v) {print "|$INfile|    |$OUTfile|     |$ALLfile|\n\n";}
-			if (!$T)
-				{
-				`mv -f "$dir1/$INfile" "$dir1/ORIG/$INfile"`;
-				`mv -f "$dir1/$OUTfile" "$dir1/ORIG/$OUTfile"`;
-				`mv -f "$dir1/$ALLfile" "$dir1/DONE/$ALLfile"`;
-				}
-			else
-				{
-				`cp -f "$dir1/$ALLfile" "$dir1/DONE/$ALLfile"`;
-				}
-
-			`$lamebin -b 16 -m m --silent "$dir1/DONE/$ALLfile" "$dir1/DONE/$MP3file"`;
-
-			if (!$T)
-				{
-				`rm -f "$dir1/DONE/$ALLfile"`;
-				}
-
-			if($DB){print STDERR "\n|/usr/bin/sox $live_folder/$filename[$k]$WAV $arch_folder/$filename[$k]$GSM|\n";}
-		chmod 0755, "$dir1/DONE/$MP3file";
-
 	### BEGIN Remote file transfer
-	#		$p = Net::Ping->new();
-	#		$ping_good = $p->ping("$FTP_host");
+			$p = Net::Ping->new();
+			$ping_good = $p->ping("$FTP_host");
 
-	#		if ($ping_good)
-	#			{
+			if ($ping_good)
+				{
+					`$soxmixbin "$dir1/$INfile" "$dir1/$OUTfile" "$dir1/$ALLfile"`;
+				if ($v) {print "|$INfile|    |$OUTfile|     |$ALLfile|\n\n";}
+
+					`$lamebin -b 16 -m m --silent "$dir1/$ALLfile" "$dir1/$MP3file"`;
+
+					if($DB){print STDERR "\n|/usr/bin/sox $live_folder/$filename[$k]$WAV $arch_folder/$filename[$k]$GSM|\n";}
+				chmod 0755, "$dir1/DONE/$MP3file";
+
 				$ftp = Net::FTP->new("$FTP_host", Port => $FTP_port, Debug => 0,  Passive => 1);
 				$ftp->login("$FTP_user","$FTP_pass");
 				$ftp->cwd("$FTP_dir");
 				$ftp->binary();
 				$ftp->put("$dir1/DONE/$MP3file", "$MP3file");
 				$ftp->quit;
-	#			}
+
+				if (!$T)
+					{
+					`mv -f "$dir1/$INfile" "$dir1/ORIG/$INfile"`;
+					`mv -f "$dir1/$OUTfile" "$dir1/ORIG/$OUTfile"`;
+					`mv -f "$dir1/$ALLfile" "$dir1/DONE/$ALLfile"`;
+					}
+				else
+					{
+					`cp -f "$dir1/$ALLfile" "$dir1/DONE/$ALLfile"`;
+					}
+				if (!$T)
+					{
+					`rm -f "$dir1/DONE/$ALLfile"`;
+					}
+
+				}
 	### END Remote file transfer
 
 			}
